Rename taxon-helpers to taxonomy-data-helpers
Offers a slightly better hint about what this file contains.
